Add pair comparison feature and UI enhancements
This update includes the addition of pair comparison feature which allows for the comparison of new pair distribution data to old pair distribution data. The UI has also been updated to support the new feature with added labels, text panes, and buttons. The detailed console display and participant statistical view have been streamlined for more effective use. Additionally, translations for the new features have to be added, everything is in English currently